Move `types` and `utils` outside `useManageSession`
