+++ frontend/src/session/types.tsx
@@ -1,4 +1,4 @@
-import { MessageContent } from "../../types";
+import { MessageContent } from "../types";
 
 export interface SessionInfo {
   id: string;
